requirements.txt
biopython>=1.83
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
xgboost>=2.0
lightgbm>=4.2
torch>=2.1
fair-esm>=2.0
matplotlib>=3.8
seaborn>=0.13
shap>=0.44
joblib>=1.3
scipy>=1.12
imbalanced-learn>=0.12
tqdm>=4.66
